MULE-10139: Move NamedObject, NameableObject and AnnotatedObject to mule-api
