+++ core-tests/src/test/java/org/mule/runtime/core/api/LocatedMuleExceptionTestCase.java
@@ -13,6 +13,8 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.mockito.Mockito.withSettings;
 
+import org.mule.runtime.api.meta.AnnotatedObject;
+import org.mule.runtime.api.meta.NamedObject;
 import org.mule.tck.junit4.AbstractMuleContextTestCase;
 import org.mule.tck.size.SmallTest;
