+++ core/src/main/java/org/mule/runtime/core/api/EncryptionStrategy.java
@@ -6,6 +6,7 @@
  */
 package org.mule.runtime.core.api;
 
+import org.mule.runtime.api.meta.NamedObject;
 import org.mule.runtime.core.api.lifecycle.Initialisable;
 import org.mule.runtime.core.api.security.CryptoFailureException;
