+++ core/src/main/java/org/mule/runtime/core/exception/MessagingExceptionLocationProvider.java
@@ -9,7 +9,7 @@ package org.mule.runtime.core.exception;
 import static org.mule.runtime.core.api.LocatedMuleException.INFO_LOCATION_KEY;
 import static org.mule.runtime.core.api.LocatedMuleException.INFO_SOURCE_XML_KEY;
 
-import org.mule.runtime.core.api.AnnotatedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
 import org.mule.runtime.core.api.MuleEvent;
 import org.mule.runtime.core.api.construct.MessageProcessorPathResolver;
 import org.mule.runtime.core.api.execution.LocationExecutionContextProvider;
