+++ core/src/main/java/org/mule/runtime/core/lifecycle/processor/ProcessIfStartedMessageProcessor.java
@@ -7,7 +7,7 @@
 package org.mule.runtime.core.lifecycle.processor;
 
 import org.mule.runtime.core.api.MuleEvent;
-import org.mule.runtime.core.api.NameableObject;
+import org.mule.runtime.api.meta.NameableObject;
 import org.mule.runtime.core.api.lifecycle.LifecycleException;
 import org.mule.runtime.core.api.lifecycle.LifecycleState;
 import org.mule.runtime.core.api.lifecycle.Startable;
