+++ modules/db/src/main/java/org/mule/runtime/module/db/internal/config/domain/database/DbConfigResolverFactoryBean.java
@@ -7,7 +7,7 @@
 
 package org.mule.runtime.module.db.internal.config.domain.database;
 
-import org.mule.runtime.core.api.AnnotatedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
 import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.context.MuleContextAware;
 import org.mule.runtime.core.api.lifecycle.Disposable;
