+++ modules/db/src/main/java/org/mule/runtime/module/db/internal/domain/database/DbConfig.java
@@ -7,8 +7,8 @@
 
 package org.mule.runtime.module.db.internal.domain.database;
 
-import org.mule.runtime.core.api.AnnotatedObject;
-import org.mule.runtime.core.api.NamedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
+import org.mule.runtime.api.meta.NamedObject;
 import org.mule.common.Testable;
 import org.mule.common.metadata.ConnectorMetaDataEnabled;
 import org.mule.runtime.module.db.internal.domain.connection.DbConnectionFactory;
