+++ modules/http/src/main/java/org/mule/runtime/module/http/api/listener/HttpListenerConfig.java
@@ -6,9 +6,9 @@
  */
 package org.mule.runtime.module.http.api.listener;
 
-import org.mule.runtime.core.api.AnnotatedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
 import org.mule.runtime.core.api.MuleException;
-import org.mule.runtime.core.api.NamedObject;
+import org.mule.runtime.api.meta.NamedObject;
 import org.mule.runtime.core.api.config.Config;
 import org.mule.runtime.core.api.lifecycle.Startable;
 import org.mule.runtime.core.api.lifecycle.Stoppable;
