+++ modules/http/src/main/java/org/mule/runtime/module/http/api/requester/HttpRequesterConfig.java
@@ -6,8 +6,8 @@
  */
 package org.mule.runtime.module.http.api.requester;
 
-import org.mule.runtime.core.api.AnnotatedObject;
-import org.mule.runtime.core.api.NamedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
+import org.mule.runtime.api.meta.NamedObject;
 import org.mule.runtime.core.api.lifecycle.Startable;
 import org.mule.runtime.core.api.lifecycle.Stoppable;
 import org.mule.runtime.module.http.api.HttpAuthentication;
