+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/factories/WatermarkFactoryBean.java
@@ -7,7 +7,7 @@
 
 package org.mule.runtime.config.spring.factories;
 
-import org.mule.runtime.core.api.AnnotatedObject;
+import org.mule.runtime.api.meta.AnnotatedObject;
 import org.mule.runtime.core.api.MuleContext;
 import org.mule.runtime.core.api.context.MuleContextAware;
 import org.mule.runtime.core.api.store.ObjectStore;
