+++ modules/spring-config/src/main/java/org/mule/runtime/config/spring/processors/MuleObjectNameProcessor.java
@@ -6,7 +6,7 @@
  */
 package org.mule.runtime.config.spring.processors;
 
-import org.mule.runtime.core.api.NameableObject;
+import org.mule.runtime.api.meta.NameableObject;
 import org.mule.runtime.core.api.agent.Agent;
 import org.mule.runtime.core.api.transformer.Transformer;
 import org.mule.runtime.core.api.transport.LegacyConnector;
