+++ tests/integration-tranports/src/test/java/org/mule/test/config/EndpointTransformerTestCase.java
@@ -13,7 +13,7 @@ import org.mule.compatibility.core.api.endpoint.EndpointFactory;
 import org.mule.compatibility.core.api.endpoint.ImmutableEndpoint;
 import org.mule.functional.transformer.NoActionTransformer;
 import org.mule.runtime.core.api.MuleException;
-import org.mule.runtime.core.api.NameableObject;
+import org.mule.runtime.api.meta.NameableObject;
 import org.mule.tck.junit4.AbstractMuleContextEndpointTestCase;
 
 import org.junit.Test;
