+++ tests/integration-tranports/src/test/java/org/mule/test/integration/config/ResponseTransformerTestCase.java
@@ -11,7 +11,7 @@ import static org.junit.Assert.assertFalse;
 
 import org.mule.compatibility.core.api.endpoint.InboundEndpoint;
 import org.mule.functional.junit4.FunctionalTestCase;
-import org.mule.runtime.core.api.NameableObject;
+import org.mule.runtime.api.meta.NameableObject;
 import org.mule.runtime.core.api.processor.MessageProcessor;
 import org.mule.runtime.core.construct.Flow;
 import org.mule.runtime.core.processor.chain.InterceptingChainLifecycleWrapper;
